move relayBundle to core
